+++ lib/scoped.dart
@@ -2,9 +2,9 @@ library scoped;
 
 export 'package:scoped/src/store.dart';
 export 'package:scoped/src/scope.dart';
-export 'package:scoped/src/fluid.dart';
+export 'package:scoped/src/reactive.dart';
 export 'package:scoped/src/ref.dart';
-export 'package:scoped/src/strip.dart';
+export 'package:scoped/src/refs.dart';
 
 export 'package:scoped/src/bond.dart';