+++ lib/src/fluids.dart
@@ -1,34 +1,34 @@
-import 'package:scoped/src/fluid.dart';
+import 'package:scoped/src/reactive.dart';
 import 'package:flutter/widgets.dart';
 
-// Fluid builder with multiple fluids
-typedef Widget FluidsWidgetBuilder(BuildContext context, List<Fluid> states);
+// Fluid builder with multiple reactive models
+typedef Widget ReactivesWidgetBuilder(BuildContext context, List<Reactive> models);
 
-class FluidsBuilder extends StatefulWidget {
-  final List<Fluid> states;
-  final FluidsWidgetBuilder builder;
-  FluidsBuilder({this.states, this.builder});
-  _FluidsBuilderState createState() => new _FluidsBuilderState();
+class ReactivesBuilder extends StatefulWidget {
+  final List<Reactive> states;
+  final ReactivesWidgetBuilder builder;
+  ReactivesBuilder({this.states, this.builder});
+  _ReactivesBuilderState createState() => new _ReactivesBuilderState();
 }
 
-class _FluidsBuilderState<T extends Fluid> extends State<FluidsBuilder> {
+class _ReactivesBuilderState<T extends Reactive> extends State<ReactivesBuilder> {
   void _onChange() => setState(() {});
 
   void initState() {
     super.initState();
-    widget.states.forEach((f) => f.listen(_onChange));
+    widget.states.forEach((f) => f.addListener(_onChange));
   }
 
   void dispose() {
-    widget.states.forEach((f) => f.forget(_onChange));
+    widget.states.forEach((f) => f.removeListener(_onChange));
     super.dispose();
   }
 
-  void didUpdateWidget(FluidsBuilder oldWidget) {
+  void didUpdateWidget(ReactivesBuilder oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.states != oldWidget.states) {
-      oldWidget.states.forEach((f) => f.forget(_onChange));
-      widget.states.forEach((f) => f.listen(_onChange));
+      oldWidget.states.forEach((f) => f.removeListener(_onChange));
+      widget.states.forEach((f) => f.addListener(_onChange));
     }
   }